myproject/requirements_test.txt
# Testing requirements for Django project
# Install with: pip install -r requirements_test.txt

# Core testing frameworks
pytest>=7.4.0
pytest-django>=4.5.2
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-xdist>=3.3.1

# Coverage reporting
coverage>=7.2.0
pytest-html>=3.2.0

# HTML parsing for template tests
beautifulsoup4>=4.12.2

# Performance testing
pytest-benchmark>=4.0.0

# Async testing support
pytest-asyncio>=0.21.0

# Fixtures and factories
factory-boy>=3.3.0

# Mock and testing utilities
responses>=0.23.0
freezegun>=1.2.0

# Linting and code quality (optional)
flake8>=6.0.0
black>=23.0.0
isort>=5.12.0

# Security testing
safety>=2.3.0
bandit>=1.7.0

# Documentation testing
pytest-doctestplus>=1.0.0
